Fix CSS. Fix bugs relating to page.
Made it so that the time till air does not split across two lines on longer anime titles.
Round the time till airing.
Added a check for url path on dom change to ensure user is still on animelist, as well as for when it could not find a list element.